fix(sdk-c): segfault potential in Keeper client (#599)
* fix(sdk-c): segfault potential in Keeper client

The delayed message-bus connect thread could remain running when
the service is exiting. If the service or service->msgbus pointers
were non-null but pointed to freed memory, segfaults could ensue.

The fix: add a flag to exit the thread, ensure it is set when
calling edgex_keeper_client_free(), and the thread is waited on
in service.c before proceeding with the exit.

Also move the "stopconfig" flag check so the sleep(1) is not
between that check and the subscription attempt.

Closes: #598

Signed-off-by: Corey Mutter <CoreyMutter@eaton.com>

* fix: rework to use 'stopconfig' as the only needed stop variable

Signed-off-by: Corey Mutter <CoreyMutter@eaton.com>

---------

Signed-off-by: Corey Mutter <CoreyMutter@eaton.com>